--- a/builder/bdwgc.go
+++ b/builder/bdwgc.go
@@ -14,7 +14,7 @@ var BoehmGC = Library{
 	name: "bdwgc",
 	cflags: func(target, headerPath string) []string {
 		libdir := filepath.Join(goenv.Get("TINYGOROOT"), "lib/bdwgc")
-		return []string{
+		flags := []string{
 			// use a modern environment
 			"-DUSE_MMAP",              // mmap is available
 			"-DUSE_MUNMAP",            // return memory to the OS using munmap
@@ -30,6 +30,7 @@ var BoehmGC = Library{
 			// Use a minimal environment.
 			"-DNO_MSGBOX_ON_ERROR", // don't call MessageBoxA on Windows
 			"-DDONT_USE_ATEXIT",
+			"-DNO_GETENV",
 
 			// Special flag to work around the lack of __data_start in ld.lld.
 			// TODO: try to fix this in LLVM/lld directly so we don't have to
@@ -49,12 +50,13 @@ var BoehmGC = Library{
 
 			"-I" + libdir + "/include",
 		}
+		return flags
 	},
 	needsLibc: true,
 	sourceDir: func() string {
 		return filepath.Join(goenv.Get("TINYGOROOT"), "lib/bdwgc")
 	},
-	librarySources: func(target string) ([]string, error) {
+	librarySources: func(target string, _ bool) ([]string, error) {
 		sources := []string{
 			"allchblk.c",
 			"alloc.c",