--- a/compileopts/config.go
+++ b/compileopts/config.go
@@ -22,7 +22,8 @@ import (
 // builder.Library struct but that's hard to do since we want to know the
 // library path in advance in several places).
 var libVersions = map[string]int{
-	"musl": 3,
+	"musl":  3,
+	"bdwgc": 2,
 }
 
 // Config keeps all configuration affecting the build in a single struct.
@@ -133,7 +134,7 @@ func (c *Config) GC() string {
 // that can be traced by the garbage collector.
 func (c *Config) NeedsStackObjects() bool {
 	switch c.GC() {
-	case "conservative", "custom", "precise":
+	case "conservative", "custom", "precise", "boehm":
 		for _, tag := range c.BuildTags() {
 			if tag == "tinygo.wasm" {
 				return true
@@ -258,6 +259,15 @@ func MuslArchitecture(triple string) string {
 	return CanonicalArchName(triple)
 }
 
+// Returns true if the libc needs to include malloc, for the libcs where this
+// matters.
+func (c *Config) LibcNeedsMalloc() bool {
+	if c.GC() == "boehm" && c.Target.Libc == "wasi-libc" {
+		return true
+	}
+	return false
+}
+
 // LibraryPath returns the path to the library build directory. The path will be
 // a library path in the cache directory (which might not yet be built).
 func (c *Config) LibraryPath(name string) string {
@@ -281,9 +291,14 @@ func (c *Config) LibraryPath(name string) string {
 		archname += "-v" + strconv.Itoa(v)
 	}
 
+	options := ""
+	if c.LibcNeedsMalloc() {
+		options += "+malloc"
+	}
+
 	// No precompiled library found. Determine the path name that will be used
 	// in the build cache.
-	return filepath.Join(goenv.Get("GOCACHE"), name+"-"+archname)
+	return filepath.Join(goenv.Get("GOCACHE"), name+options+"-"+archname)
 }
 
 // DefaultBinaryExtension returns the default extension for binaries, such as