--- a/src/runtime/arch_tinygowasm_malloc.go
+++ b/src/runtime/arch_tinygowasm_malloc.go
@@ -1,4 +1,4 @@
-//go:build tinygo.wasm && !(custommalloc || wasm_unknown)
+//go:build tinygo.wasm && !(custommalloc || wasm_unknown || gc.boehm)
 
 package runtime