--- a/targets/wasip1.json
+++ b/targets/wasip1.json
@@ -23,7 +23,8 @@
 		"--no-demangle"
 	],
 	"extra-files": [
-		"src/runtime/asm_tinygowasm.S"
+		"src/runtime/asm_tinygowasm.S",
+		"src/runtime/gc_boehm.c"
 	],
 	"emulator":      "wasmtime run --dir={tmpDir}::/tmp {}"
 }