See: https://github.com/tinygo-org/tinygo/pull/4385

This is an incomplete port. Things that need to be resolved are:

 - [x] #4385 needs to be merged first, and this PR needs to be rebased.
 - [x] We need to have two separate builds of wasi-libc since the `MALLOC_IMPL` macro is going to be different. I think it's easiest to start doing this build not when creating a TinyGo tarball but when running TinyGo itself (similar to picolibc, musl, etc). This makes it easier to build different variants easily and when needed.

The performance improvements can be large. For example, for the html package some benchmarks are more than 50% faster:

```
$ benchstat html-precise.txt html-boehm.txt
               │ html-precise.txt │           html-boehm.txt           │
               │      sec/op      │   sec/op     vs base               │
Escape               110.29µ ± 0%   73.62µ ± 3%  -33.25% (p=0.002 n=6)
EscapeNone            48.37µ ± 0%   49.15µ ± 1%   +1.62% (p=0.002 n=6)
Unescape             192.31µ ± 5%   84.53µ ± 3%  -56.04% (p=0.002 n=6)
UnescapeNone          2.638µ ± 1%   2.559µ ± 4%        ~ (p=0.132 n=6)
UnescapeSparse        45.44µ ± 2%   18.55µ ± 1%  -59.17% (p=0.002 n=6)
UnescapeDense        182.55µ ± 1%   79.63µ ± 1%  -56.38% (p=0.002 n=6)
geomean               53.11µ        32.40µ       -39.00%
```

Pause times should also be reduced quite a bit, likely much more than the performance improvements. I did not test those, though.

If you want to try this PR, you will have to rebuild wasi-libc:

```
$ rm -rf lib/wasi-libc/sysroot/ lib/wasi-libc/build
$ make wasi-libc
```

(Also note that only wasip1 is supported right now, other wasm variants should be easy to add in the future).