@aykevl 2025-04-09
The main drawback right now is that it will often output messages like the following:

```
GC Warning: Repeated allocation of very large block (appr. size 68 KiB):
        May lead to memory leak and poor performance
```

I think the only reasonable fix is to just disable output entirely (see https://github.com/ivmai/bdwgc/pull/721) or disable it using `GC_set_warn_proc` since using `GC_malloc_ignore_off_page` is not safe with Go (it's allowed to take the pointer of an object, so `GC_malloc_ignore_off_page` is unsafe even if it would probably work for most programs in practice).