@deadprogram 2025-04-22
@aykevl now that https://github.com/ivmai/bdwgc/commit/6d10870ba8ca4410276af90c7fc305f1d2949c00 was merged, I think that this PR can proceed?