@aykevl 2025-05-23
Updated the PR.
* rebased on dev
* added `GC_DISABLE_SNPRINTF=1` on WebAssembly to avoid linking in printf related functions (saves ~18kB of binary size)
* disabled warnings on WebAssembly since they happen frequently and there's not much we can do about them.