@deadprogram 2025-05-23
@aykevl looks like actual error here: https://github.com/tinygo-org/tinygo/actions/runs/15212794972/job/42791925530?pr=4812#step:17:24