@aykevl 2025-05-23
@deadprogram actually, you might be right. The Boehm GC is also used on Windows, and this PR updates the bdwgc library.