@aykevl 2025-05-30
I think it's related to https://github.com/ivmai/bdwgc/issues/736. Probably our bug and not one in bdwgc, but asking to be sure.