@deadprogram 2025-06-02
@aykevl not quite passing yet https://github.com/tinygo-org/tinygo/actions/runs/15388465010/job/43292271205#step:17:24