@deadprogram 2025-06-02
OK, new error https://github.com/tinygo-org/tinygo/actions/runs/15389453659/job/43296579801?pr=4812#step:7:28