@aykevl 2025-06-11
I'm not so sure about that. I think you just managed to avoid a somewhat rare crash this time, but the bug is still there.