@deadprogram 2025-06-12
> I'm not so sure about that. I think you just managed to avoid a somewhat rare crash this time, but the bug is still there.

@aykevl I don't think that is specific to this PR, do you?