@aykevl 2025-06-13
Changed to use the bdwgc commit just before https://github.com/ivmai/bdwgc/commit/e8aedf635e1ad42102799478ea8a2ab55abd2ba5 which I think introduced the Windows issues (GOARCH=386 at least, but I also suspect it is causing the flaky CI issue). See https://github.com/ivmai/bdwgc/issues/736#issuecomment-2929578337 for details. This does mean we lose some of the recent size improvements, but fixing these issues will need a bit more time. To at least have (opt-in) bdwgc support in wasm for this release (even if it grows the binary more than needed) is better than having nothing.