@deadprogram 2025-06-13
> To at least have (opt-in) bdwgc support in wasm for this release (even if it grows the binary more than needed) is better than having nothing.

Agreed!